ENH: automatically update `pyright` hook (#326)
* MAINT: write unit tests for `pyright` module

+++ src/compwa_policy/check_dev_files/pyright.py
@@ -4,19 +4,34 @@ from __future__ import annotations
 
 import json
 import os
+from pathlib import Path
+from typing import TYPE_CHECKING
 
-from compwa_policy.utilities.pyproject import ModifiablePyproject, complies_with_subset
+from compwa_policy.utilities.precommit.struct import Hook, Repo
+from compwa_policy.utilities.pyproject import (
+    ModifiablePyproject,
+    Pyproject,
+    complies_with_subset,
+)
 from compwa_policy.utilities.toml import to_toml_array
 
+if TYPE_CHECKING:
+    from compwa_policy.utilities.precommit import ModifiablePrecommit
 
-def main() -> None:
+
+def main(precommit: ModifiablePrecommit) -> None:
     with ModifiablePyproject.load() as pyproject:
         _merge_config_into_pyproject(pyproject)
+        _update_precommit(precommit, pyproject)
         _update_settings(pyproject)
 
 
-def _merge_config_into_pyproject(pyproject: ModifiablePyproject) -> None:
-    old_config_path = "pyrightconfig.json"  # cspell:ignore pyrightconfig
+def _merge_config_into_pyproject(
+    pyproject: ModifiablePyproject,
+    path: Path = Path("pyrightconfig.json"),
+    remove: bool = True,
+) -> None:
+    old_config_path = path
     if not os.path.exists(old_config_path):
         return
     with open(old_config_path) as stream:
@@ -26,14 +41,25 @@ def _merge_config_into_pyproject(pyproject: ModifiablePyproject) -> None:
             existing_config[key] = to_toml_array(sorted(value))
     tool_table = pyproject.get_table("tool.pyright", create=True)
     tool_table.update(existing_config)
-    os.remove(old_config_path)
+    if remove:
+        os.remove(old_config_path)
     msg = f"Imported pyright configuration from {old_config_path}"
     pyproject.append_to_changelog(msg)
 
 
+def _update_precommit(precommit: ModifiablePrecommit, pyproject: Pyproject) -> None:
+    if not __has_pyright(pyproject):
+        return
+    repo = Repo(
+        repo="https://github.com/ComPWA/mirrors-pyright",
+        rev="",
+        hooks=[Hook(id="pyright")],
+    )
+    precommit.update_single_hook_repo(repo)
+
+
 def _update_settings(pyproject: ModifiablePyproject) -> None:
-    table_key = "tool.pyright"
-    if not pyproject.has_table(table_key):
+    if not __has_pyright(pyproject):
         return
     pyright_settings = pyproject.get_table("tool.pyright")
     minimal_settings = {
@@ -43,3 +69,8 @@ def _update_settings(pyproject: ModifiablePyproject) -> None:
         pyright_settings.update(minimal_settings)
         msg = "Updated pyright configuration"
         pyproject.append_to_changelog(msg)
+
+
+def __has_pyright(pyproject: Pyproject) -> bool:
+    table_key = "tool.pyright"
+    return pyproject.has_table(table_key)
